Refactor: Extract common template methods to TemplateRequestHelpersTrait
Address the SonarQube issue regarding duplicated code. This change promotes code reusability and maintainability across the models.
