+++ ProcessMaker/Templates/ProcessTemplate.php
@@ -4,7 +4,6 @@ namespace ProcessMaker\Templates;
 
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Auth;
 use ProcessMaker\Events\TemplateCreated;
@@ -16,7 +15,6 @@ use ProcessMaker\ImportExport\Options;
 use ProcessMaker\Models\Process;
 use ProcessMaker\Models\ProcessCategory;
 use ProcessMaker\Models\ProcessTemplates;
-use ProcessMaker\Models\Template;
 use ProcessMaker\Models\WizardTemplate;
 use ProcessMaker\Traits\HasControllerAddons;
 use SebastianBergmann\CodeUnit\Exception;
@@ -28,6 +26,7 @@ use Spatie\MediaLibrary\MediaCollections\Models\Media;
 class ProcessTemplate implements TemplateInterface
 {
     use HasControllerAddons;
+    use TemplateRequestHelperTrait;
 
     const PROJECT_ASSET_MODEL_CLASS = 'ProcessMaker\Package\Projects\Models\ProjectAsset';
 
@@ -40,14 +39,13 @@ class ProcessTemplate implements TemplateInterface
      */
     public function index(Request $request)
     {
-        $orderBy = $this->getRequestSortBy($request, 'name');
-        $include = $this->getRequestInclude($request);
-        $templates = ProcessTemplates::nonSystem()->with($include);
-
-        $filter = $request->input('filter');
-
-        $templates = $templates
-            ->select(
+        return ProcessTemplates::nonSystem()
+            ->with($this->getRequestInclude($request))
+            ->withFilters($request->input('filter'))
+            ->orderBy(...$this->getRequestSortBy($request, 'name'))
+            ->leftJoin('process_categories as category', 'process_templates.process_category_id', '=', 'category.id')
+            ->leftJoin('users as user', 'process_templates.user_id', '=', 'user.id')
+            ->select([
                 'process_templates.id',
                 'process_templates.uuid',
                 'process_templates.key',
@@ -62,28 +60,8 @@ class ProcessTemplate implements TemplateInterface
                 'process_templates.asset_type',
                 'process_templates.created_at',
                 'process_templates.updated_at',
-            )
-            ->leftJoin('process_categories as category', 'process_templates.process_category_id', '=', 'category.id')
-            ->leftJoin('users as user', 'process_templates.user_id', '=', 'user.id')
-            ->orderBy(...$orderBy)
-            ->where(function ($query) use ($filter) {
-                $query->where('process_templates.name', 'like', '%' . $filter . '%')
-                    ->orWhere('process_templates.description', 'like', '%' . $filter . '%')
-                    ->orWhere('user.firstname', 'like', '%' . $filter . '%')
-                    ->orWhere('user.lastname', 'like', '%' . $filter . '%')
-                    ->orWhereIn('process_templates.id', function ($qry) use ($filter) {
-                        $qry->select('assignable_id')
-                            ->from('category_assignments')
-                            ->leftJoin('process_categories', function ($join) {
-                                $join->on('process_categories.id', '=', 'category_assignments.category_id');
-                                $join->where('category_assignments.category_type', '=', ProcessCategory::class);
-                                $join->where('category_assignments.assignable_type', '=', ProcessTemplates::class);
-                            })
-                            ->where('process_categories.name', 'like', '%' . $filter . '%');
-                    });
-            })->get();
-
-        return $templates;
+            ])
+            ->get();
     }
 
     /**
@@ -500,51 +478,6 @@ class ProcessTemplate implements TemplateInterface
         } catch (\Exception $e) {
             return response()->json(['message' => $e->getMessage()], 422);
         }
-    }
-
-    /**
-     * Get process template manifest.
-     *
-     * @param string $type
-     *
-     * @param Request $request
-     *
-     * @return array
-     */
-    public function getManifest(string $type, int $id) : array
-    {
-        $response = (new ExportController)->manifest($type, $id);
-
-        return json_decode($response->getContent(), true);
-    }
-
-    /**
-     * Get included relationships.
-     *
-     * @param Request $request
-     *
-     * @return array
-     */
-    protected function getRequestSortBy(Request $request, $default) : array
-    {
-        $column = $request->input('order_by', $default);
-        $direction = $request->input('order_direction', 'asc');
-
-        return [$column, $direction];
-    }
-
-    /**
-     * Get included relationships.
-     *
-     * @param Request $request
-     *
-     * @return array
-     */
-    protected function getRequestInclude(Request $request) : array
-    {
-        $include = $request->input('include');
-
-        return $include ? explode(',', $include) : [];
     }
 
     /**
